[APPS]Enable exFAT and regenerated hub_msd demo for pic32mx470_curiosity configuration.[MH3-84718]

+++ apps/hub_msd/firmware/src/config/pic32mx470_curiosity/system/fs/sys_fs.h
@@ -661,7 +661,7 @@ typedef void (* SYS_FS_EVENT_HANDLER)
 typedef struct 
 {
     /* File size */
-    uint32_t    fsize;
+    uint64_t    fsize;
     /* Last modified date */
     uint16_t    fdate;
     /* Last modified time */
